// types/meta.ts
export interface MetaData {
  title?: string
  description?: string
  keywords?: string
  author?: string
  favicon?: string
  charset?: string
  viewport?: string
  robots?: string
  generator?: string
  theme?: string
  canonical?: string
  language?: string
  ogTitle?: string
  ogDescription?: string
  ogImage?: string
  twitterCard?: string
  twitterTitle?: string
  twitterDescription?: string
  twitterImage?: string
  icons?: string[]
  [key: string]: any
}